requirements.txt
streamlit==1.21.0
streamlit-option-menu==0.3.2
plotly==5.14.1
pandas==1.5.3
Pillow==9.5.0